+++ packages/cli/services/trialDatastore/grpcservers/trialDatastoreServer_test.go
@@ -26,6 +26,7 @@ import (
 	grpcapi "github.com/cogment/cogment/grpcapi/cogment/api"
 	"github.com/cogment/cogment/services/trialDatastore/backend"
 	"github.com/cogment/cogment/services/trialDatastore/backend/memoryBackend"
+	"github.com/cogment/cogment/services/utils"
 	"github.com/stretchr/testify/assert"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
@@ -43,7 +44,7 @@ type trialDatastoreServerTestFixture struct {
 
 func createTrialDatastoreServerTestFixture() (trialDatastoreServerTestFixture, error) {
 	listener := bufconn.Listen(1024 * 1024)
-	server := CreateGrpcServer(false)
+	server := utils.NewGrpcServer(false)
 	backend, err := memoryBackend.CreateMemoryBackend(memoryBackend.DefaultMaxSampleSize)
 	if err != nil {
 		return trialDatastoreServerTestFixture{}, err